+++ examples/readme_example.rs
@@ -42,8 +42,8 @@ fn main() {
                     },
                     //Other input messages here
                 },
-                NetEvent::AddedEndpoint(_endpoint) => println!("TCP Client connected"),
-                NetEvent::RemovedEndpoint(_endpoint) => println!("TCP Client disconnected"),
+                NetEvent::Connected(_endpoint) => println!("TCP Client connected"),
+                NetEvent::Disconnected(_endpoint) => println!("TCP Client disconnected"),
                 NetEvent::DeserializationError(_) => (),
             },
             // Other events here