+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - Fixed some memory leaks at decoding pool.
 - Removed unused `Network::local_addr`.
 - Renamed from `SendingStatus` to `SendStatus`.
+- Renamed from `MAX_UDP_LEN` to `MAX_UDP_PAYLOAD_LEN`.
+- Renamed from `AddedEndpoint` to `Connected`.
+- Renamed from `RemovedEndpoint` to `Disconnected`.
 
 ## Release 0.6.0
 - Added concurrent writing and reading from socket/streams in UDP and TCP protocols.